+++ CHANGES.md
@@ -13,3 +13,4 @@
 * added filter and function to remove slugs from appearing for updates
 * added a Settings tab to add site data, removed JSON storage directory
 * added nonce checks for WP_List_Table
+* move Git Remote Updater a top level menu